chore: renamed controllers & adjust methods
