--- src/main/java/com/distinction/barbenheimer/service/SeatPurchaseService.java
@@ -1,7 +1,0 @@
-package com.distinction.barbenheimer.service;
-
-
-public interface SeatPurchaseService {
-
-        public String createCustomerIdentifyingToken();
-}
